--- CurrencyLayer/CurrencyLayer/Common/Views/StatefulView/EmptyScreen.swift
@@ -1,14 +1,0 @@
-//
-//  EmptyScreen.swift
-//  CurrencyLayer
-//
-//  Created by Filipe Cruz on 14/12/20.
-//
-import SwiftUI
-
-struct EmptyScreen:  View {
-  var body: some View {
-    Text("No Itens Found")
-      .frame(maxWidth: .infinity, alignment: .center)
-  }
-}
